src/views: new AddRule view
